+++ e2e/list-toolbar.e2e-spec.ts
@@ -10,12 +10,23 @@ import {
 
 describe('list-toolbar component', () => {
 
-  function validateScreenshot(done: DoneFn) {
-    expect('#screenshot-list-toolbar').toMatchBaselineScreenshot(done);
-  }
-
   beforeEach(() => {
     SkyHostBrowser.get('visual/list-toolbar');
+    SkyHostBrowser.setWindowBreakpoint('lg');
+  });
+
+  it('should display toolbar correctly', (done) => {
+    expect('#screenshot-list-toolbar').toMatchBaselineScreenshot(done, {
+      screenshotName: 'list-toolbar'
+    });
+  });
+
+  it('should display toolbar correctly (xs screen)', (done) => {
+    SkyHostBrowser.setWindowBreakpoint('xs');
+
+    expect('#screenshot-list-toolbar').toMatchBaselineScreenshot(done, {
+      screenshotName: 'list-toolbar-xs'
+    });
   });
 
   it('should display toolbar with the column chooser', (done) => {
@@ -23,22 +34,49 @@ describe('list-toolbar component', () => {
       by.css('#screenshot-list-toolbar .sky-list-secondary-actions .sky-dropdown-button')
     ).click();
 
-    expect('#screenshot-list-toolbar').toMatchBaselineScreenshot(done);
+    expect('#screenshot-list-toolbar').toMatchBaselineScreenshot(done, {
+      screenshotName: 'list-toolbar-with-column-chooser'
+    });
+  });
+
+  it('should display toolbar with the column chooser (xs screen)', (done) => {
+    SkyHostBrowser.setWindowBreakpoint('xs');
+    element(
+      by.css('#screenshot-list-toolbar .sky-list-secondary-actions .sky-dropdown-button')
+    ).click();
+
+    expect('#screenshot-list-toolbar').toMatchBaselineScreenshot(done, {
+      screenshotName: 'list-toolbar-with-column-chooser-xs'
+    });
   });
 
   it('should display inline filters correctly when opened', (done) => {
     element(by.css('#screenshot-list-toolbar .sky-filter-btn')).click();
 
-    validateScreenshot(done);
+    expect('#screenshot-list-toolbar').toMatchBaselineScreenshot(done, {
+      screenshotName: 'list-toolbar-with-inline-filters'
+    });
   });
 
-  it('should display toolbar correctly when a small screen', (done) => {
+  it('should display inline filters correctly when opened (xs screen)', (done) => {
     SkyHostBrowser.setWindowBreakpoint('xs');
+    element(by.css('#screenshot-list-toolbar .sky-filter-btn')).click();
 
-    element(
-      by.css('#screenshot-list-toolbar .sky-list-secondary-actions .sky-dropdown-button')
-    ).click();
+    expect('#screenshot-list-toolbar').toMatchBaselineScreenshot(done, {
+      screenshotName: 'list-toolbar-with-inline-filters-xs'
+    });
+  });
 
-    validateScreenshot(done);
+  it('should display multiselect action bar', (done) => {
+    expect('#screenshot-list-toolbar-with-multiselect').toMatchBaselineScreenshot(done, {
+      screenshotName: 'list-toolbar-with-multiselect'
+    });
+  });
+
+  it('should display multiselect action bar (xs screen)', (done) => {
+    SkyHostBrowser.setWindowBreakpoint('xs');
+    expect('#screenshot-list-toolbar-with-multiselect').toMatchBaselineScreenshot(done, {
+      screenshotName: 'list-toolbar-with-multiselect'
+    });
   });
 });